services/grpc_server/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
googleapis-common-protos==1.57.0
pymongo==4.3.3